move ValidityView into its own module, like ChunkedView is
