+++ js/src/vector/view.ts
@@ -1,8 +1,9 @@
 export { ChunkedView } from './chunked';
+export { ValidityView } from './validity';
 export { DictionaryView } from './dictionary';
 export { ListView, FixedSizeListView, BinaryView, Utf8View } from './list';
 export { UnionView, DenseUnionView, NestedView, StructView, MapView } from './nested';
-export { FlatView, NullView, BoolView, ValidityView, PrimitiveView, FixedSizeView, Float16View } from './flat';
+export { FlatView, NullView, BoolView, PrimitiveView, FixedSizeView, Float16View } from './flat';
 export { DateDayView, DateMillisecondView } from './flat';
 export { IntervalYearMonthView, IntervalYearView, IntervalMonthView } from './flat';
 export { TimestampDayView, TimestampSecondView, TimestampMillisecondView, TimestampMicrosecondView, TimestampNanosecondView } from './flat';
